Source Change: Fix `::owner()` and `::group()`.
Add stricter check for `$owneruid` and `$gid`.

Previously, a falsy check led to `::owner()` and `::group()`
returning `false` for a file owned by the root user, whose ID is `0`.

This has now been changed to a strict check for accuracy.

For falsy checks on the result of `::owner()` and `::group()`,
there will be no change.

For strict checks on the result of `::owner()` and `::group()`,
the result will be more accurate.